refactor:CatAnimationの歩行モーション分離、型定義
